Use i32::MIN for mouse coordinate sentinels
Initialize mouse_x and mouse_y with i32::MIN instead of casting i8::MIN to i32. This ensures the full i32 sentinel range is used (avoiding narrowing/truncation) in the SDL2 platform implementation (crates/pyxel-core/src/platform/sdl2/platform_sdl2.rs).

+++ crates/pyxel-core/src/platform/sdl2/platform_sdl2.rs
@@ -57,8 +57,8 @@ impl PlatformSdl2 {
             window: null_mut(),
             gl_context: null_mut(),
             audio_device_id: 0,
-            mouse_x: i8::MIN as i32,
-            mouse_y: i8::MIN as i32,
+            mouse_x: i32::MIN,
+            mouse_y: i32::MIN,
             is_wayland: false,
             gamepads: Vec::new(),
             #[cfg(target_os = "emscripten")]
